introduces helpers for common blocks of nodeadm e2e tests
